requirements.txt
numpy==1.23.3
python-dotenv==1.0.1
pandas==1.5.0
pycryptodome==3.16.0
requests==2.28.1
setuptools==65.6.3
websocket_client==1.4.2
asyncio
schedule